feat(ux): Streamline scanner flow & fix pre-commit violations
- Removed redundant 'Select Payslip Photo' screen (immediate gallery access)
- Removed obsolete 'Auto/Officer/JCO/OR' format selector
- Preserved first-time privacy education flow
- Refactored PDFProcessingService to extract JCO/OR helpers to extension
- Fixed file length violation in PDFProcessingService.swift (<300 lines)
- Verified clean build and tests (JCOORRoutingTests passed)
