+++ PayslipMax/Views/Payslips/PayslipScannerView.swift
@@ -2,8 +2,8 @@ import SwiftUI
 import VisionKit
 import UIKit
 
-/// Full-screen photo selector that routes captured images through the payslip processing pipeline.
-/// Uses WhatsApp-style gallery-first approach with privacy education.
+/// Streamlined photo selector that opens gallery immediately.
+/// Privacy education shown once for first-time users, then direct gallery access.
 struct PayslipScannerView: View {
     @Environment(\.dismiss) private var dismiss
 
@@ -14,10 +14,9 @@ struct PayslipScannerView: View {
     @State private var isProcessing = false
     @State private var errorMessage: String?
     @State private var showError = false
-    @State private var showingPhotoPicker = true // Show gallery-first
+    @State private var showingPhotoPicker = false
     @State private var showingCropper = false
     @State private var pendingImage: UIImage?
-    @State private var userHint: PayslipUserHint = .auto
     @State private var showPrivacyEducation = false
 
     init(onFinished: (() -> Void)? = nil, onImageCaptured: ((UIImage) -> Void)? = nil) {
@@ -26,53 +25,12 @@ struct PayslipScannerView: View {
     }
 
     var body: some View {
-        NavigationStack {
-            ZStack {
-                // Background
-                Color(.systemGroupedBackground)
-                    .ignoresSafeArea()
-
-                VStack(spacing: 0) {
-                    // Privacy tip banner for returning users
-                    if UserDefaults.hasSeenPrivacyEducation {
-                        privacyTipBanner
-                            .padding(.horizontal)
-                            .padding(.top, 8)
-                    }
-
-                    // Hint selector
-                    hintSelector
-                        .padding(.horizontal, 16)
-                        .padding(.top, 12)
-
-                    Spacer()
-
-                    // Instructions
-                    instructionsView
-
-                    Spacer()
-                }
-
-                if isProcessing {
-                    Color.black.opacity(0.35)
-                        .ignoresSafeArea()
-                    ProgressView("Processing payslip...")
-                        .padding()
-                        .background(.ultraThinMaterial)
-                        .cornerRadius(10)
-                }
-            }
-            .navigationTitle("Select Payslip Photo")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button("Cancel") {
-                        dismiss()
-                    }
-                }
-            }
+        // No intermediate screen - gallery opens immediately
+        Color.clear
             .alert("Error", isPresented: $showError) {
-                Button("OK", role: .cancel) { }
+                Button("OK", role: .cancel) {
+                    dismiss()
+                }
             } message: {
                 Text(errorMessage ?? "An unknown error occurred.")
             }
@@ -114,17 +72,33 @@ struct PayslipScannerView: View {
             .sheet(isPresented: $showPrivacyEducation) {
                 PrivacyEducationSheet {
                     showPrivacyEducation = false
-                    showingPhotoPicker = true
+                    // Small delay to ensure smooth sheet transition
+                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
+                        showingPhotoPicker = true
+                    }
                 }
             }
             .onAppear {
-                // Show privacy education for first-time users
+                // First-time users: Show privacy education, then gallery
+                // Returning users: Open gallery immediately
                 if !UserDefaults.hasSeenPrivacyEducation {
                     showPrivacyEducation = true
-                    showingPhotoPicker = false
+                } else {
+                    showingPhotoPicker = true
                 }
             }
-        }
+            .overlay {
+                if isProcessing {
+                    ZStack {
+                        Color.black.opacity(0.35)
+                            .ignoresSafeArea()
+                        ProgressView("Processing payslip...")
+                            .padding()
+                            .background(.ultraThinMaterial)
+                            .cornerRadius(10)
+                    }
+                }
+            }
     }
 
     // MARK: - Private
@@ -141,7 +115,8 @@ struct PayslipScannerView: View {
     }
 
     private func processCroppedImageWithLLM(_ image: UIImage) async {
-        imageProcessor.updateUserHint(userHint)
+        // LLM handles format detection automatically - no user hint needed
+        imageProcessor.updateUserHint(.auto)
         let result = await imageProcessor.processCroppedImageLLMOnly(image)
 
         switch result {
@@ -162,60 +137,5 @@ struct PayslipScannerView: View {
             showError = true
             isProcessing = false
         }
-    }
-
-    // MARK: - UI Components
-
-    private var privacyTipBanner: some View {
-        HStack(spacing: 8) {
-            Image(systemName: "info.circle.fill")
-                .foregroundColor(.blue)
-                .font(.footnote)
-
-            Text("Next: Crop personal info before AI scan")
-                .font(.footnote)
-                .foregroundColor(.secondary)
-
-            Spacer()
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 8)
-        .background(Color.blue.opacity(0.1))
-        .cornerRadius(8)
-    }
-
-    private var hintSelector: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Text("Parsing preference")
-                .font(.caption)
-                .foregroundColor(.secondary)
-                .accessibilityHidden(true)
-            Picker("Parsing preference", selection: $userHint) {
-                Text("Auto").tag(PayslipUserHint.auto)
-                Text("Officer").tag(PayslipUserHint.officer)
-                Text("JCO/OR").tag(PayslipUserHint.jcoOr)
-            }
-            .pickerStyle(.segmented)
-            .accessibilityIdentifier("payslip_hint_picker")
-        }
-    }
-
-    private var instructionsView: some View {
-        VStack(spacing: 16) {
-            Image(systemName: "photo.on.rectangle.angled")
-                .font(.system(size: 60))
-                .foregroundColor(.secondary)
-
-            Text("Select a payslip photo from your gallery")
-                .font(.headline)
-                .multilineTextAlignment(.center)
-                .foregroundColor(.primary)
-
-            Text("You'll crop out sensitive info in the next step")
-                .font(.subheadline)
-                .multilineTextAlignment(.center)
-                .foregroundColor(.secondary)
-        }
-        .padding(.horizontal, 32)
     }
 }
